feat(ios): 약관 관련 내용 정리
