+++ Sources/TossPayments/UIKit/Widget/CanUpdateHeight.swift
@@ -1,0 +1,12 @@
+//
+//  CanUpdateHeight.swift
+//  
+//
+//  Created by 김진규 on 2023/05/31.
+//
+
+import Foundation
+
+public protocol CanUpdateHeight: AnyObject {
+    var updatedHeight: CGFloat { get set }
+}
