--- Sources/TossPayments/UIKit/Widget/PaymentWidget+WKUIDelegate.swift
@@ -1,25 +1,0 @@
-//
-//  PaymentWidget+WKUIDelegate.swift
-//
-//
-//  Created by https://github.com/MussaCharles on 2023/04/19.
-//
-
-import WebKit
-
-// MARK: - WKUIDelegate
-
-extension PaymentMethodWidget: WKUIDelegate, BrowserPopupHandler {
-    
-    public func webView(
-        _ webView: WKWebView,
-        createWebViewWith configuration: WKWebViewConfiguration, 
-        for navigationAction: WKNavigationAction, 
-        windowFeatures: WKWindowFeatures
-    ) -> WKWebView? {
-        let popupWebView = createPopupWindow(parentWebView: webView, configuration: configuration)
-        present(popupWebView: popupWebView)
-        return popupWebView
-        
-    }
-}
